move examples and delete tkinter
